Sass reconfiguration using chokidar
