+++ nwb.config.js
@@ -3,14 +3,5 @@ module.exports = {
   npm: {
     esModules: true,
     umd: false
-  },
-  webpack: {
-    rules: {
-      'sass-css': {
-        modules: {
-          localIdentName: '[name]__[local]__[hash:base64:5]'
-        }
-      }
-    }
   }
 }
